examples/example_1.py
# main.py

import abc

import numpy as np
import yaml

from configurable import TypedConfigurable, Schema


# Component Definitions

class DataPreprocessor(TypedConfigurable, abc.ABC):
    aliases = ['data_preprocessor']

    @abc.abstractmethod
    def preprocess(self, data):
        pass

class NormalizingPreprocessor(DataPreprocessor):
    aliases = ['normalizing_preprocessor']
    config_schema = {
        'normalization': Schema(bool, default=True),
        'resize': Schema(int, default=256),
    }

    def preprocess(self, data):
        if self.normalization:
            data = (data - data.mean()) / data.std()
        return data.resize((self.resize, self.resize))

class Model(TypedConfigurable, abc.ABC):
    aliases = ['model']

    @abc.abstractmethod
    def train(self, data):
        pass

class AdvancedModel(Model):
    aliases = ['advanced_model']
    config_schema = {
        'layers': Schema(int, default=50),
        'dropout': Schema(float, default=0.5),
    }

    def train(self, data):
        # Simulate training process
        return f"Training with {self.layers} layers and {self.dropout} dropout"

class Optimizer(TypedConfigurable, abc.ABC):
    aliases = ['optimizer']

    @abc.abstractmethod
    def optimize(self, model):
        pass

class AdamOptimizer(Optimizer):
    aliases = ['adam_optimizer']
    config_schema = {
        'learning_rate': Schema(float, default=0.001),
    }

    def optimize(self, model):
        return f"Optimizing with learning rate {self.learning_rate}"

# Pipeline Configuration

class AIPipeline:
    def __init__(self, config_path):
        with open(config_path, 'r') as file:
            config = yaml.safe_load(file)

        self.data_preprocessor = DataPreprocessor.from_config(config['pipeline']['data_preprocessor'])
        self.model = Model.from_config(config['pipeline']['model'])
        self.optimizer = Optimizer.from_config(config['pipeline']['optimizer'])

    def run(self, data):
        preprocessed_data = self.data_preprocessor.preprocess(data)
        training_result = self.model.train(preprocessed_data)
        optimization_result = self.optimizer.optimize(self.model)
        return training_result, optimization_result

# YAML Configuration (embedded as a string for a single file)

config_yaml = """
pipeline:
  data_preprocessor:
    type: 'normalizing_preprocessor'
    normalization: true
    resize: 256
  model:
    type: 'advanced_model'
    layers: 50
    dropout: 0.5
  optimizer:
    type: 'adam_optimizer'
    learning_rate: 0.001
"""

# Pipeline Execution

if __name__ == "__main__":
    # Simulate some data
    data = np.random.rand(100, 100, 3)

    # Write the YAML configuration to a temporary file
    with open('config.yaml', 'w') as file:
        file.write(config_yaml)

    # Initialize and run the pipeline
    pipeline = AIPipeline('config.yaml')
    training_result, optimization_result = pipeline.run(data)

    print("Training Result:", training_result)
    print("Optimization Result:", optimization_result)
